Lab1 - in the process

+++ src/ru/chigurov/lab1/Main.java
@@ -3,34 +3,37 @@ package ru.chigurov.lab1;
 import java.util.Scanner;
 
 public class Main {
-
+    /**
+     *  программа работает от аргумента, в зависимости от вида представления(чай или кофе)
+     */
     public static void main(String[] args) {
-	// write your code here
 
-     Tea tea = new Tea();
-        Scanner scanner = new Scanner(System.in);
-        System.out.println("Выберите действие:\n" +
-                "1. Создание объекта\n" +
-                "2. Вывод объекта на экран\n" +
-                "3. Ввод данных с клавиатуры\n" +
-                "4. Принудительное зануление данных\n" +
-                "5. Выход");
-        int menuSelection = scanner.nextInt();
-        while (menuSelection != 5){
-            if (menuSelection == 1){
-                tea.create();
+        if (args[0].equalsIgnoreCase("tea")) {
+            Tea tea = new Tea();
+            Scanner scanner = new Scanner(System.in);
+            System.out.println("Выберите действие:\n" +
+                    "1. Создание объекта\n" +
+                    "2. Вывод объекта на экран\n" +
+                    "3. Ввод данных с клавиатуры\n" +
+                    "4. Принудительное зануление данных\n" +
+                    "5. Выход");
+            int menuSelection = scanner.nextInt();
+            while (menuSelection != 5) {
+                if (menuSelection == 1) {
+                    tea.create();
+                }
+                if (menuSelection == 2) {
+                    tea.read();
+                }
+                if (menuSelection == 3) {
+                    tea.update();
+                }
+                if (menuSelection == 4) {
+                    tea.delete();
+                }
+                menuSelection = scanner.nextInt();
             }
-            if (menuSelection == 2){
-                tea.read();
-            }
-            if (menuSelection == 3){
-                tea.update();
-            }
-            if (menuSelection == 4){
-                tea.delete();
-            }
-            menuSelection = scanner.nextInt();
+
         }
-
     }
 }
